Fix build for go < 1.18
